✨ Feat : 로컬 환경 적용 & .env 적용
