+++ src/Page/Home/HomeImageBarrel.ts
@@ -4,8 +4,9 @@ import Banner from '../../image/Home/Banner.png';
 import Logo_002 from '../../image/Home/Logo_002.png';
 import 헤드셋칩스 from '../../image/Home/헤드셋칩스.png';
 import 기본프로필 from '../../image/예시사진모음/default_profile_W.png';
+import Share_Icon from '../../image/Icon/Share_Icon.svg';
 
-export { chip_Active, chip_NoneActive, Banner, Logo_002, 헤드셋칩스, 기본프로필 };
+export { chip_Active, chip_NoneActive, Banner, Logo_002, 헤드셋칩스, 기본프로필, Share_Icon };
 
 // export * from '../../image/Home/chip_Active.png';
 // export * from '../../image/Home/chip_NoneActive.png';
